requirements.txt
fastapi==0.115.12
uvicorn==0.23.2
face-recognition==1.3.0
python-socketio==5.10.0
python-dotenv==1.0.0
requests==2.32.0
Pillow==10.1.0
python-multipart==0.0.20

# For the face recognition model
git+https://github.com/ageitgey/face_recognition_models

numpy==2.2.5
setuptools==80.0.1